Add generic substitution list
This is a new settings option, users can even edit it in the
settings UI without touching settings.json.
It is a mapping of strings to strings that will be run through
.replaceAll before speaking the text.

I made voice only be a string type so the user can edit it through the
settings UI.

Please note that VS Code automatically updated tasks.json
when I opened this project.
"Task version 0.1.0 is deprecated. Please use 2.0.0"
On the other hand, I updated typescript to 4.0.0 so I could use .replaceAll.
I changed @types/vscode to 1.5.0 to match the engines.vscode value
so vsce wouldn't complain.